# specs/001-banking-crud/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Banking Microservices API
  description: |
    RESTful API for banking operations including customer management,
    account management, and transaction processing.

    ## Authentication
    All endpoints except `/auth/login` and `/auth/register` require a valid JWT token
    in the Authorization header: `Bearer <token>`

    ## Rate Limiting
    - Authentication endpoints: 5 requests/minute
    - Read operations: 100 requests/minute
    - Write operations: 30 requests/minute
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.example.com/v1
    description: Production
...
paths:
...
  /accounts/{account_number}/transactions:
    parameters:
      - name: account_number
        in: path
        required: true
        schema:
          type: string
          pattern: '^\d{16}$'

    get:
      tags: [Transactions]
      summary: Get transaction history
      operationId: listTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: Filter transactions from this date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: Filter transactions until this date
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
          description: Filter by transaction type
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Number of transactions to return
      responses:
        '200':
          description: Transaction history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Transactions]
      summary: Create a transaction (deposit or withdrawal)
      operationId: createTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Validation error or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/{account_number}/transactions/{reference}:
    parameters:
      - name: account_number
        in: path
        required: true
        schema:
          type: string
          pattern: '^\d{16}$'
      - name: reference
        in: path
        required: true
        schema:
          type: string
          pattern: '^TXN[A-Z0-9]+$'

    get:
      tags: [Transactions]
      summary: Get transaction details
      operationId: getTransaction
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # HEALTH & METRICS
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransactionType:
      type: string
      enum: [deposit, withdrawal]

    # ========== AUTH ==========
...
    TransactionCreate:
      type: object
      required: [transaction_type, amount]
      properties:
        transaction_type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
          format: decimal
          exclusiveMinimum: 0
          example: 100.00
        description:
          type: string
          maxLength: 255
    TransactionResponse:
      type: object
      properties:
        reference:
          type: string
          example: TXN20260116ABC123
        transaction_type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
          format: decimal
        balance_after:
          type: number
          format: decimal
        description:
          type: string
        created_at:
          type: string
          format: date-time
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        total:
          type: integer
        next_cursor:
          type: string
          description: Cursor for next page, null if no more results

    # ========== ERROR ==========
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
              additionalProperties: true
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
              details: {}
    Forbidden:
      description: Access denied to this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to access this resource
              details: {}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found
              details: {}
...